Implement type completion based on directory structure
This commit implements the method `ede-php-autoload-complete-type-name`
for each class loader. This method takes the beginning of a fully
qualified name as input, and returns completion suggestions on it.

For PSR-0, it also infers whether the type is written with underscores
or antislashes.

This method is then used to implement
`semanticdb-find-tags-for-completion-method` to get auto-completion in
the semanticdb backend.

refs #5